백엔드 리팩토링

+++ models/project.js
@@ -1,6 +1,6 @@
 
 const mysql = require('mysql2/promise')
-const settings = require('../config/appsettings.local.json')
+const settings = require('../modules/config')
 
 // TODO: 정규식
 
@@ -15,12 +15,8 @@ class Project {
         try {
             connection = await this._pool.getConnection()
 
-            const [rows] = await connection.query(`SELECT * FROM project WHERE id=${pid}`)
-
-            if (rows.length == 0)
-                return null
-
-            return rows[0]
+            const [[row]] = await connection.query(`SELECT * FROM project WHERE id=${pid}`)
+            return row
         } catch (err) {
             throw err
         } finally {
@@ -34,9 +30,8 @@ class Project {
         try {
             connection = await this._pool.getConnection()
 
-            const [rows] = await connection.query(`SELECT * FROM project WHERE owner=${user_id}`)
-
-            return rows
+            const [[row]] = await connection.query(`SELECT * FROM project WHERE owner=${user_id}`)
+            return row
         } catch (err) {
             throw err
         } finally {
@@ -71,12 +66,12 @@ class Project {
 
             // TODO: 프로젝트 이름 정규식
             // 자신이 만든 프로젝트에 중복된 팀 이름이 있는지 검사
-            const [exists] = await connection.query(
+            const [[exists]] = await connection.query(
                 `SELECT COUNT(*) AS count FROM project
                 WHERE name='${project.name}' AND owner=${user_id}`
             )
 
-            if (exists[0].count > 0)
+            if (exists.count > 0)
                 return null
 
             const [result] = await connection.query(
@@ -96,28 +91,11 @@ class Project {
         try {
             connection = await this._pool.getConnection()
 
-            // project의 관리자 찾기
-            const [project_owner] = await connection.query(
-                `SELECT owner FROM project WHERE id=${project_id}`
-            )
-            
-            if(project_owner[0].owner != user_id)
-                throw "관리자가 아닙니다."
-
-            const [exists] = await connection.query(
-                `SELECT COUNT(*) AS count FROM project
-                WHERE NOT id=${project_id} AND owner=${user_id} AND name='${project.name}'`
-            )
-
-            if (exists[0].count > 0)
-                return null
-            
             const [result] = await this._pool.query(
-                `UPDATE project SET name='${project.name}' WHERE id=${project_id}`
+                `UPDATE project SET name=${project.name} WHERE id=${project_id} AND owner=${user_id}`
             )
 
             return result.affectedRows > 0
-
         } catch (err) {
             throw err
         } finally {
@@ -130,17 +108,8 @@ class Project {
 
         try {
             connection = await this._pool.getConnection()
-
-            // team의 관리자 찾기
-            const [project_owner] = await connection.query(
-                `SELECT owner FROM project WHERE id=${project_id}`
-            )
-            
-            if(project_owner[0].owner != user_id)
-                throw "관리자가 아닙니다."
-
             const [result] = await connection.query(
-                `DELETE FROM project WHERE id=${project_id}`
+                `DELETE FROM project WHERE id=${project_id} AND owner=${user_id}`
             )
 
             return result.affectedRows > 0
